+++ tests/tests_pytorch/deprecated_api/test_remove_1-7.py
@@ -20,7 +20,6 @@ import pytest
 import torch
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.demos.boring_classes import BoringModel
 from pytorch_lightning.plugins.environments import (
     KubeflowEnvironment,
     LightningEnvironment,
@@ -37,21 +36,6 @@ def test_v1_7_0_deprecate_lightning_distributed(tmpdir):
         from pytorch_lightning.distributed.dist import LightningDistributed
 
         _ = LightningDistributed()
-
-
-def test_v1_7_0_deprecate_on_post_move_to_device(tmpdir):
-    class TestModel(BoringModel):
-        def on_post_move_to_device(self):
-            print("on_post_move_to_device")
-
-    model = TestModel()
-
-    trainer = Trainer(default_root_dir=tmpdir, limit_train_batches=5, max_epochs=1)
-
-    with pytest.deprecated_call(
-        match=r"Method `on_post_move_to_device` has been deprecated in v1.5 and will be removed in v1.7"
-    ):
-        trainer.fit(model)
 
 
 def test_v1_7_0_deprecated_max_steps_none(tmpdir):